+++ packages/application-store/src/workouts/get-workouts-selector.ts
@@ -3,6 +3,6 @@ import { ApplicationState } from '../store';
 export const getWorkoutsSelector = (state: ApplicationState) => state.workouts;
 export const getWorkoutById = (state: ApplicationState, workoutId:string) => {
     const workout = state.workouts.find(({ id }) => id === workoutId);
-    if(!workout) return state.workouts[0]
+    if(!workout) return undefined;
     return workout;
 };
