+++ packages/liftograph-app/src/pages/liftograph-main/liftograph-main.tsx
@@ -1,11 +1,9 @@
-import { Box, ListItem, ListItemText, Tab, Tabs } from '@mui/material';
+import { Box, Tab, Tabs } from '@mui/material';
 import React, { useState } from 'react';
-import { v4 } from 'uuid';
 import TabPanel from '../../components/tab-panel';
 import styles from './styles.less'
 import { Link } from 'react-router-dom';
-import { Button } from '@mui/base';
-import { addWorkout, getWorkoutsSelector, useStoreDispatch, useStoreSelection } from '@liftograph/application-store';
+import { WorkoutDisplay } from '../../components/workout';
 
 export const WORKOUT_EDITOR_LINK_DATA_TESTID = 'workout-editor-link';
 export const TAB_DATA_TESTID = 'liftograph-tab-';
@@ -17,8 +15,6 @@ export enum TabApplication {
 
 export function LiftographMain() {
     const [currentTab, setCurrentTab] = useState(TabApplication.Workouts);
-    const workouts = useStoreSelection(getWorkoutsSelector);
-    const dispatch = useStoreDispatch();
 
     return (
         <div className={styles.liftographTabContainer}>
@@ -58,37 +54,7 @@ export function LiftographMain() {
 
                     {/* Below is temporary to test state is working. */}
                     <br />
-                    <Button
-                        onClick={() => {
-                            const name = prompt('Name your workout') ?? 'test';
-                            dispatch(
-                                addWorkout({
-                                    name,
-                                    id: v4(),
-                                    exercises: []
-                                })
-                            );
-                        }}
-                    >
-                        Add Workout
-                    </Button>
-
-                    <h3>Workouts</h3>
-                    {
-                        workouts.map((workout) => {
-                            return (
-                                <ListItem>
-                                    <Link
-                                        data-testid={WORKOUT_EDITOR_LINK_DATA_TESTID}
-                                        to='workout-editor'
-                                        state={{workoutId: workout.id}}
-                                    >
-                                        <ListItemText primary={workout.name}/>
-                                    </Link>
-                                </ListItem>
-                            );
-                        })
-                    }
+                    <WorkoutDisplay/>
                 </TabPanel>
                 <TabPanel
                     currentTab={currentTab}
